debt - move fileSearchManager to common (for #84106)
